+++ datamodel/libCore/tests/Reference_test.cpp
@@ -10,6 +10,8 @@
 #include "core/BasicTypes.h"
 #include "data_storage/Value.h"
 
+#include "testing/TestUtil.h"
+
 #include "user_types/Node.h"
 #include "user_types/MeshNode.h"
 #include "user_types/Mesh.h"
@@ -118,11 +120,11 @@ TEST(ReferenceTest, Property) {
 	SMeshNode meshnode_b{new MeshNode("meshnode b")};
 	SMesh mesh{new Mesh("mesh")};
 
-	Property<SNode> vnode = node;
-	Property<SNode> vnode_b = node_b;
-	Property<SMeshNode> vmeshnode = meshnode;
-	Property<SMeshNode> vmeshnode_b = meshnode_b;
-	Property<SMesh> vmesh = mesh;
+	Property<SNode, Dummy> vnode{node, {}};
+	Property<SNode, Dummy> vnode_b{node_b, {}};
+	Property<SMeshNode, Dummy> vmeshnode{meshnode, {}};
+	Property<SMeshNode, Dummy> vmeshnode_b{meshnode_b, {}};
+	Property<SMesh, Dummy> vmesh{mesh, {}};
 
 	// Assigning shared_ptr<T> to Property<U>s
 
@@ -145,7 +147,7 @@ TEST(ReferenceTest, Property) {
 	EXPECT_EQ(*vnode, node_b);
 
 	// assign child class
-	vnode = vmeshnode;
+	vnode = *vmeshnode;
 	EXPECT_EQ(*vnode, meshnode);
 
 	// assign unrelated class: doesn't compile
@@ -173,12 +175,12 @@ TEST(ReferenceTest, TypeEquality) {
 	SNode node2{new Node("node2")};
 	SMeshNode meshnode{new MeshNode("meshnode")};
 
-	Property<SNode> pnode = node;
-	Property<SNode> pnode2 = node2;
+	Property<SNode, Dummy> pnode{node, {}};
+	Property<SNode, Dummy> pnode2{node2, {}};
 	Value<SNode> vnode = node;
 	Property<SNode, RangeAnnotation<double>> pnoderange{node, RangeAnnotation<double>()};
 	Property<SNode, HiddenProperty> pnodehidden {node, HiddenProperty()};
-	Property<SMeshNode> pmeshnode = meshnode;
+	Property<SMeshNode, Dummy> pmeshnode{meshnode, {}};
 
 	EXPECT_TRUE(ValueBase::classesEqual(pnode, pnode2));
 	EXPECT_FALSE(ValueBase::classesEqual(vnode, pnode));